requirements.txt
pandas>=2.0.0
pyspark>=3.4.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.11.0
anthropic>=0.25.0
google-genai>=1.0.0
openai>=1.0.0
pyyaml>=6.0
jsonschema>=4.18.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
flask>=3.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
numpy>=1.25.0
scipy>=1.11.0
